feat: apply changes from Webview to Document
The editor in webview can change through DevTools, for example.

We use `MutationObserver` to watch for DOM changes on webview, and send
mutations to the extention.

We can't send `Node` directly without a serializer, so we calculate its
XPath instead. The extension uses that XPath to apply the mutations to
the document model.

The editor inside Webview is implemented as a web component that moves
the body content into shadow DOM and manipulates it there.

Even though the backend can know the exactt line and column number of
the mutation target, we still apply edit to the whole document.