feat: SSI as portal
The embeded SSI content is converted into a custom element that acts like
the new `<portal>`.

Content is kept in light DOM, but wrapped in a `<a>` inside shadow DOM.

Scoped styles are used to prevent user interaction on content.

The portal has `src` of vscode's command URI to open the target in a new
Tailwind editor.