refactor: portals for all
Instead of using a single `vscode-portal` element, which may break the
styles of the element it represents, we create one custom portal foreach
HTML elements.

In order to avoid bloat, we iterate through the `portal-*` elements on
the page and constructs custom portal for them.

By using `is` attribute instead, the HTML markup remain the same as the
template, but it acts as a portal.